requirements.txt
Django>=5.1
django-otp>=1.0
djangorestframework>=3.15
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
requests>=2.31
python-dotenv>=1.0
whitenoise>=6.5
gunicorn>=22.0
django-filter>=24.0
openpyxl==3.1.5
reportlab==4.4.10
